move support related container to subdir
