+++ src/containers/support/SupportContactUs.tsx
@@ -1,8 +1,8 @@
 import { faGithub } from '@fortawesome/free-brands-svg-icons';
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import * as React from "react";
-import PageHeader from "../components/PageHeader";
-import SupportConfirmation from '../content/supportConfirmation.mdx';
+import PageHeader from "../../components/PageHeader";
+import SupportConfirmation from '../../content/supportConfirmation.mdx';
 
 import SupportContactUsForm from './SupportContactUsForm';
